plugins/rosetta-prompt/rosetta_prompt/requirements.txt
# LangChain v1 
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# API Framework
fastapi>=0.115.0
uvicorn>=0.32.0

# Data validation
pydantic>=2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
